+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for service version 2019-12-12.
 - Added support for Jumbo Files.
 - Fixed bug where DataLakeFileClient, DataLakeDirectoryClient, and DataLakePathClient.Name and .Path were sometimes URL-encoded.
+- Fixed bug where DataLakeDirectoryClient.GetSubDirectory(), GetFile(), CreateSubDirectory(), and CreateFile() were returning clients with an incorrect URI.
 
 ## 12.2.2 (2020-06)
 - This release contains bug fixes to improve quality.